Expanding IWireCoil API
